OCR/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
paddleocr==2.10.0
paddlepaddle-gpu==3.0.0b1   # or paddlepaddle==3.0.0b1 if using CPU only
spacy==3.8.6
spacy-transformers==1.3.8
torch==2.6.0+cu118
transformers==4.49.0
opencv-python==4.11.0.86
pandas==2.2.3
numpy==2.2.5
scikit-image==0.25.2
scipy==1.15.2
matplotlib==3.10.1

# Optional: SpaCy models
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl
